Standardize CNI context
